Fix the query manager's memory watchdog.
This actually makes sure that memory limits are abided by. This was
a serious bug of the revised query API. Now, the memory
limiting logic is applied as close to the processing as possible.

The DB work manager will listen for "out-of-allowed-memory" messages
and stop it's individual workers. To be able to do that, the workers
were converted into stateful goroutines that can be canceled at any
given time during workload processing.