[compression] null-compress when appropriate
Writes the raw bytes to file in case the compressed size of the data
is bigger than the raw data itself.

This makes compression slightly slower due to compressed data that isn't
used in the end. However, storage is saved and decompression speed is increased,
saving decompression cycles where appropriate