Fix/simplify sub-command handling and flag parsing
Removes the need for sub-command handling inside the Execute() method
of the rootCmd, relying on the spf13's default way of building up the
command tree.

To make this work, the Args function inside the rootCmd had to be
overwritten to not complain when a query such as "sip,dip" was passed
as a command. See `validatePositionalArgs` in cmd/goQuery/cmd/root.go
for details.

Other changes:
* make sure that packets come before bytes in the list output to be
  consistent with the normal query output
* remove the --list flag. It's now exclusively handled by the "list"
  command

TODO:
* make sure the fringes of the time range are respected (partial time
  frames). For a full time frame the results are consistent with the
  iface query:

    goQuery -i any -f '-1000000d' iface