Handle partial days (tfirst and tlast)
This will handle cases where the list target doesn't span the entire
DB, but only a subset of it (e.g. --first or --last are specified and
fall into the days in which data was captured).

Refactoring:
* to simplify the walking of the DB (and make the code re-usable) the loops
around year, month, day have been re-arranged in a walkDB() function of
the DBWorkManager.

Functionality:
* the log level can now be configured with flag --logging.level. It is set to
  Warn per default. This is employed for debugging in the ReadMetadata function

Open issues:
* time frames that fall into the same day will lead to a panic in some cases
  (see TODO in DBWorkManager.go)
* if the first time frame concerns the last day of available data, curDir is not
  closed and goQuery blocks forever