Instrument the server(s) with pprof
Provides facilities to hook up a net/http/pprof server on /debug/pprof.

The majority of this commit will unify the server base across all binaries,
which serve an API, in this case `goProbe` and `global-query`. This is primarily
to deduplicate code for functionalit that is desired in both APIs.

To this end, and analogous to the client.DefaultClient a server.DefaultServer
is introduced to cover observability middlewares as well as setting up the gin.Engin