[feature] Add flow map volume/packets metrics to goProbe and introduce iface label (#237)
### Background

This commits add three new metrics to goProbe:

* `goprobe_capture_bytes_total`
* `goprobe_capture_packets_total`
* `goprobe_capture_flows_total`

These are calculated and supplied to prometheus only upon aggregation,
which is in line with the approach chosen for the other capture metrics
during `status()` calls. The `bytes` and `packets` metrics also have a
label `direction=(inbount|outbound)`.

### Major Changes

Let's unleash cardinality onto the metrics backends. Metrics of the
`capture` subsystem are now vectors with exactly one label: `iface`.

The decision was made to discern better

* on which interface errors/drops occured
* how large the flow maps were
* how much traffic volume was observed per interface

The latter two providing a more detailed understanding of bursts or
interfaces over which many sessions are routed.

Closes #236

---------

Co-authored-by: Lennart Elsen <lelsen@open-systems.com>