[bugfix] Remove cyclic calls for Prettier / Pretty() interface & implementation (#275)
This gets rid of the crashes. However TBH I'm not supa happy with the
current setup, mostly because it's inconsistent. There are places where
it's asserted if an error is a `Prettier`, then a few lines later an
error is generated that just assumes to be. We should revisit the whole
console output / pretty / non-pretty concept IMHO at some point (and be
as generic as possible, adding a few helper functions maybe to make our
lives easier).