[feature] Switch API server to `huma` for OpenAPI docs auto generation/documentation (#311)
Simplifies the API documentation massively, as it is now embedded in the
struct tags. Closes #309 .

# API

Under the hood, the [huma](https://huma.rocks/) framework is used:

- documents API via struct tags in relevant entities
- replaces route registration from the `gin` framework

Not only does this simplify the hooks into the API, it also doesn't
touch the remaining server setup fundamentally. The router in front of
`huma` is still `gin`.

The API documentation is reachable under the `/docs` endpoint.

# Error Handling

This addresses parts of #277. The detail errors for parsing args now
relies on the `huma.ErrorModel`, which returns an RFC-compliant list of
errors.

The benefit is that every field in the query args is validated and the
whole list of offending fields is returned in the error. Both the API
and `goQuery` use this to provide details to the caller.

For `goQuery`, the output looks as follows

```shell
$ go run cmd/goQuery/main.go -i eth0 sip,dap -c "dport 89" -n 0
Error running query: failed to prepare query:

    query (value: sip,dap)
    ----------------------
    invalid query type: 
    
    sip,dap
        ^
    unknown attribute
    
    condition (value: dport 89)
    ---------------------------
    invalid condition: 
    
    dport 89
          ^
    expected comparison operator
    
    num_results (value: 0)
    ----------------------
    invalid number of result rows: must be > 0
```

# Flags / OpenAPI Generation

The OpenAPI spec can now be generated by the tool itself. This is
necessary in order to guarantee up-to-date generation during the release
step.

To this end, the `--openapi` flag was added to both `goProbe` and
`global-query`. It causes the tools to start and generate the API spec
as soon as possible.

Since neither tools start without config, the `.buildconfig` directory
contains a minimal, stripped-down config, which can be relied on for
OpenAPI spec generation.

---------

Co-authored-by: Lennart Elsen <lelsen@open-systems.com>