[feature] Support for pluggable host resolvers in global-query (#394)
Enhances the plugin concept by allowing the host resolution
(`hosts.Resolver`) to be injected via a plugin.

The default resolver simply takes a comma-separated list of host IDs and
distributes the query to said hosts.

It is supplied by default as the `string` resolver. Also, the `static`
resolver is implemented, which takes the list of hosts to query from a
config file.

This commit makes it possible for users to define their own resolvers.
This may include:

- hosts resolution via other APIs / DSL
- resolution via DB
- discovery of hosts via discovery mechanisms

On top, the `query.Args` now support specifying the resolver type in the
query. This is first and foremost relevant if a non-standard resolver is
used and a fallback to the simple, ID-based resolver is desired.

---------

Co-authored-by: Lennart Elsen <lelsen@open-systems.com>