[feature] Validation endpoint specific to distributed querying (#399)
Modifies the existing `/_query/validate` endpoint by allowing additional
validation to be injected.

The distributed API of global-query makes use of that to validate that
the `query_hosts` parameter is set. This was previously done inside the
query controller, hence returning a 500 error.

This change makes the error 422, pointing to the offending parameter
(like for the rest of the query.Args).

Closes #398

Co-authored-by: Lennart Elsen <lelsen@open-systems.com>