[bugfix] Make sure that partial results also include rows (#402)
Addresses a bug where `rows` were `null` if part of a `partialResult` in
an event stream.

The amount of rows are guarded by the `limit` parameter, as well as a
maximum of 100 rows per partial result. This is to save bandwidth when
streaming results for queries having a large limit set.

The logic for producing the final result is now consolidated and shared
between the `sse.Send` part of the query code and the finalization code.
All results manipulation happens inside `run` now.

Finally, the `distributed` package adds tests to verify basic behavior
as well as the above for SSE sends.

Closes #401 .

---------

Co-authored-by: Lennart Elsen <lelsen@open-systems.com>