[feature] Provide access to all configured resolver plugins (#405)
More dynamic and exhaustive support for resolver plugin configuration.

Also fixes a bug introduced in
https://github.com/els0r/goProbe/pull/402, where the upper bound on the
row limit for SSE routes was applied to the final result of the query
(which should of course honor the original limit).

The remainder of the PR are linter error fixes (removing unused vars and
moving the cobra commands out of `init()` into `Execute`).

Closes #404

---------

Co-authored-by: Lennart Elsen <lelsen@open-systems.com>